resiliency: composite: add blocking routes, fix route
Adds /datetime/timeRaw and /datetime/date, that route to the respective microservice without any resiliency mechanism and without using Reactor facilities, so that thread starvation scenarios are possible.
Limits the number of running tomcat threads to 4.
Maps the original /time route to /datetime/time route so that it works correctly with the gateway.